requirements.txt
# Core dependencies
typer>=0.9.0,<0.10.0
rich>=13.4.2,<14.0.0
aiohttp>=3.8.5,<4.0.0
asyncio>=3.4.3,<4.0.0
pyyaml>=6.0,<7.0
colorama>=0.4.6,<0.5.0

# DNS and networking
dnspython>=2.4.0,<3.0.0

# Web content processing
beautifulsoup4>=4.12.2,<5.0.0
httpx>=0.24.1,<0.25.0
requests>=2.31.0,<3.0.0

# For future ML/AI features
# scikit-learn>=1.3.0,<1.4.0
# tensorflow>=2.13.0,<3.0.0
# transformers>=4.35.0,<5.0.0

# Output formats
# jinja2>=3.1.2,<3.2.0  # For HTML reports
# reportlab>=4.0.4,<4.1.0  # For PDF generation

# Testing
pytest>=7.4.0,<8.0.0
pytest-asyncio>=0.21.1,<0.22.0